Add new option to fail ratCheck if binaries are found
ASF policy is to not allow binaries in releases. This adds a new
'ratFailBinaries' setting that causes ratCheck to fail if any binaries
are found. This defaults to false to maintain existing behavior,
especially since this option requires that expected binaries (e.g.
images, test files) be excluded via the excludeFilter setting.

Fixes #16